requirements.txt
boto==2.48.0
boto3==1.4.7
botocore==1.7.32
certifi==2017.7.27.1
cfe==0.0.15
chardet==3.0.4
dj-database-url==0.4.2
Django==2.1.3
django-simple-history==2.6.0
django-storages==1.6.5
docutils==0.14
gitdb2==2.0.3
GitPython==2.1.7
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
jmespath==0.9.3
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.1
PyPDF2==1.26.0
python-dateutil==2.6.1
pytz==2017.2
reportlab==3.5.12
requests==2.18.4
s3transfer==0.1.11
six==1.11.0
smmap2==2.0.3
stripe==1.68.0
urllib3==1.22
webencodings==0.5.1
xhtml2pdf==0.2.3
